pyproject.toml
[tool.poetry]
name = "fastscanner"
version = "0.1.0"
description = ""
authors = ["Daniel Fonnegra <dfonnegrag@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "2.3.0"
numpy = "2.2.4"
python-dotenv = "^1.1.0"
httpx = "^0.28.1"
pyyaml = "^6.0.2"
pyarrow="^19.0.1"
fastapi = {extras = ["standard"], version = "^0.116.1"}
exchange-calendars= "^4.10"
redis="^5.2.1"
uvloop = "^0.21.0"
nats-py = "^2.11.0"
massive = "^2.0.2"
boto3 = "^1.42.15"
polars-lts-cpu = "^1.33.1"
beautifulsoup4 = "^4.14.3"
feedparser = "^6.0.12"
selenium = "^4.43.0"
webdriver-manager = "^4.0.2"
prometheus-client = "^0.25.0"
opentelemetry-api = "^1.41.1"
opentelemetry-sdk = "^1.41.1"
opentelemetry-exporter-otlp-proto-grpc = "^1.41.1"
opentelemetry-instrumentation-fastapi = "^0.62b1"

[tool.poetry.group.dev.dependencies]
ipython = "^9.1.0"
coverage = "^7.8.0"
pytest= "^8.3.5"
pytest-asyncio= "^0.26.0"
pytest-cov = "^7.0.0"
boto3-stubs = {extras = ["s3"], version = "^1.42.15"}
pyright = "^1.1.409"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
omit = ["test_*.py"]
